+++ __tests__/index.js
@@ -1,4 +1,4 @@
-const {lint} = require('./utils')
+import {lint} from './utils/index.js'
 
 const SAFE_SCSS_EXAMPLE = `
   .Component { float: left; }