+++ browsers.js
@@ -1,7 +1,0 @@
-module.exports = [
-  '> 5%',
-  'last 2 firefox versions',
-  'last 2 chrome versions',
-  'last 2 safari versions',
-  'last 2 edge versions',
-]