+++ plugins/typography.js
@@ -1,6 +1,6 @@
-const {createVariableRule} = require('./lib/variable-rules')
+import {createVariableRule} from './lib/variable-rules.js'
 
-module.exports = createVariableRule(
+export default createVariableRule(
   'primer/typography',
   {
     'font-size': {